+++ lib/src/features/products/view/forms/form_fields.dart
@@ -81,10 +81,13 @@ class ProductCodeFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_code),
           validator: (value) => utils.FormValidation.validateNumberField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
-            ref.read(productStateNotifierProvider.notifier).setProduct(product.copyWith(code: double.parse(value!)));
+            ref
+                .read(productStateNotifierProvider.notifier)
+                .setProduct(product.copyWith(code: int.parse(value!)));
           }),
     );
   }
@@ -101,10 +104,13 @@ class ProductNameFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_name),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_names),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_names),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
-            ref.read(productStateNotifierProvider.notifier).setProduct(product.copyWith(name: value!));
+            ref
+                .read(productStateNotifierProvider.notifier)
+                .setProduct(product.copyWith(name: value!));
           }),
     );
   }
@@ -121,10 +127,13 @@ class ProductCategoryFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_category),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_names),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_names),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
-            ref.read(productStateNotifierProvider.notifier).setProduct(product.copyWith(category: value!));
+            ref
+                .read(productStateNotifierProvider.notifier)
+                .setProduct(product.copyWith(category: value!));
           }),
     );
   }
@@ -141,7 +150,8 @@ class ProductSellRetaiPriceFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_sell_retail_price),
           validator: (value) => utils.FormValidation.validateNumberField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -163,7 +173,8 @@ class ProductSellWholePriceFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_sell_whole_price),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -185,7 +196,8 @@ class ProductSellsmanCommissionFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_salesman_comission),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -207,7 +219,8 @@ class ProductInitialQuantityFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_initial_quantitiy),
           validator: (value) => utils.FormValidation.validateNumberField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -229,7 +242,8 @@ class ProductAltertWhenLessThanFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_altert_when_less_than),
           validator: (value) => utils.FormValidation.validateNumberField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -251,7 +265,8 @@ class ProductAlertWhenExceedsFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_alert_when_exceeds),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -273,10 +288,13 @@ class ProductPackageTypeFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_package_type),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_names),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_names),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
-            ref.read(productStateNotifierProvider.notifier).setProduct(product.copyWith(packageType: value!));
+            ref
+                .read(productStateNotifierProvider.notifier)
+                .setProduct(product.copyWith(packageType: value!));
           }),
     );
   }
@@ -293,7 +311,8 @@ class ProductPackageWeightFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_package_weight),
           validator: (value) => utils.FormValidation.validateNumberField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
@@ -315,7 +334,8 @@ class ProductNumItemsInsidePackageFormField extends ConsumerWidget {
           textAlign: TextAlign.center,
           decoration: formFieldDecoration(S.of(context).product_num_items_inside_package),
           validator: (value) => utils.FormValidation.validateNameField(
-              fieldValue: value, errorMessage: S.of(context).input_validation_error_message_for_numbers),
+              fieldValue: value,
+              errorMessage: S.of(context).input_validation_error_message_for_numbers),
           onSaved: (value) {
             final product = ref.read(productStateNotifierProvider).product;
             ref
